修復 usecase

+++ src/main/java/com/eastrabbit/crudtemplate/book/usecase/delete/DeleteBookUseCaseImpl.java
@@ -1,8 +1,7 @@
 package com.eastrabbit.crudtemplate.book.usecase.delete;
 
 import com.eastrabbit.crudtemplate.book.entity.BookEntity;
-import com.eastrabbit.crudtemplate.book.usecase.BookServiceException;
-import com.eastrabbit.crudtemplate.repository.BookRepository;
+import com.eastrabbit.crudtemplate.book.repository.BookRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -11,11 +10,14 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 public class DeleteBookUseCaseImpl implements DeleteBookUseCase {
-    private final BookRepository bookRepository;
-    @Override
-    public void delete(DeleteBookUseCaseInput deleteBookUseCaseInput) throws BookServiceException {
-        Optional<BookEntity> bookEntityOpt = bookRepository.findById(deleteBookRequest.pk());
-        BookEntity bookEntity = bookEntityOpt.orElseThrow(() -> new BookServiceException("book not found , pk = " + deleteBookRequest.pk()));
-        bookRepository.delete(bookEntity);
+  private final BookRepository bookRepository;
+
+  @Override
+  public void delete(DeleteBookUseCaseInput deleteBookUseCaseInput) {
+    Optional<BookEntity> bookEntityOptional = bookRepository.findById(deleteBookUseCaseInput.pk());
+    if (bookEntityOptional.isPresent()) {
+      BookEntity bookEntity = bookEntityOptional.get();
+      bookRepository.delete(bookEntity);
     }
+  }
 }
